requirements.txt
pytest
numpy
pymodm
requests
flask
matplotlib
scikit-image
scipy
pyqt5
Pillow
pytest
